API/requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
decorator==4.4.2
Django==5.1.2
django-cors-headers==4.5.0
django-environ==0.11.2
djangorestframework==3.15.2
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.4
langchain-core==0.3.12
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.136
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.1
numba==0.60.0
numpy==1.26.4
ollama==0.3.3
openai-whisper==20240930
orjson==3.10.9
packaging==24.1
pillow==11.0.0
proglog==0.1.10
propcache==0.2.0
psutil==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
PyYAML==6.0.2
regex==2024.9.11
replicate==1.0.2
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
setuptools==75.2.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websockets==13.1
yarl==1.15.5
